+++ peersim/peersim/transport/UnreliableTransport.java
@@ -103,7 +103,7 @@ public void send(Node src, Node dest, Object msg, int pid)
 }
 
 //Comment inherited from interface
-public int getLatency(Node src, Node dest)
+public long getLatency(Node src, Node dest)
 {
 	Transport t = (Transport) src.getProtocol(pid);
 	return t.getLatency(src, dest);
